+++ src/Event.ts
@@ -212,6 +212,10 @@ export interface PointMoveEvent extends Event {
 	 * @param エンティティID
 	 */
 	10?: number;
+	/**
+	 * @param ボタンの種類
+	 */
+	11?: number;
 }
 
 export const enum PointMoveEventIndex {
@@ -225,7 +229,8 @@ export const enum PointMoveEventIndex {
 	StartDeltaY = 7,
 	PrevDeltaX = 8,
 	PrevDeltaY = 9,
-	EntityId = 10
+	EntityId = 10,
+	Button = 11,
 }
 
 /**
@@ -265,6 +270,10 @@ export interface PointUpEvent extends Event {
 	 * @param エンティティID
 	 */
 	10?: number;
+	/**
+	 * @param ボタンの種類
+	 */
+	11?: number;
 }
 
 export const enum PointUpEventIndex {
@@ -278,7 +287,8 @@ export const enum PointUpEventIndex {
 	StartDeltaY = 7,
 	PrevDeltaX = 8,
 	PrevDeltaY = 9,
-	EntityId = 10
+	EntityId = 10,
+	Button = 11
 }
 
 /**